refactor: split docs export graph
